+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2020.2 - 2020/05/02
+- Added destin, destout, destatop and srcatop
+- Updated license with credits
+- Added credits to readme (these were previously in the source file)
+
 ## 2020.1 - 2020/05/01
 - Added pinlight, vividlight and exclusion
 - Fix hardlight, negation